get rid of circ deps
